pyproject.toml
[project]
name = "fraud-risk-platform"
version = "0.1.0"
description = "Fraud risk monitoring workflow with ML scoring, threshold tuning, and analyst dashboard."
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "joblib>=1.3",
  "streamlit>=1.33",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "httpx>=0.27",
  "pillow>=10.0",
]

[project.optional-dependencies]
benchmark = [
  "xgboost>=2.0",
  "shap>=0.45",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
